# packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: Tokento API
  version: 0.1.0
  description: Tokento V1 API contract for merchant issuance, wallet query, validation, redemption, and dashboard operations.
servers:
  - url: http://localhost:4000
    description: Local development server
...
paths:
...
  /api/v1/merchants:
    post:
      tags:
        - Merchants
      summary: Create merchant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                email:
                  type: string
                  format: email
                  pattern: ^(?!\.)(?!.*\.\.)([A-Za-z0-9_'+\-\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
                settlementType:
                  default: discount
                  type: string
                  enum:
                    - discount
                    - cashback
                    - credit
                settlementTiming:
                  default: real_time
                  type: string
                  enum:
                    - real_time
                    - daily_batch
              required:
                - name
                - email
      responses:
        "201":
          description: Merchant created
          content:
            application/json:
              schema:
                type: object
                properties:
                  merchant:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      name:
                        type: string
                      email:
                        type: string
                        format: email
                        pattern: ^(?!\.)(?!.*\.\.)([A-Za-z0-9_'+\-\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
                      agentOptIn:
                        type: boolean
                      settlementType:
                        type: string
                      settlementTiming:
                        type: string
                      webhookUrl:
                        anyOf:
                          - type: string
                          - type: "null"
                      webhookSecret:
                        anyOf:
                          - type: string
                          - type: "null"
                      isSandbox:
                        type: boolean
                      createdAt:
                        type: string
                        format: date-time
                        pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                      updatedAt:
                        type: string
                        format: date-time
                        pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                    required:
                      - id
                      - name
                      - email
                      - agentOptIn
                      - settlementType
                      - settlementTiming
                      - webhookUrl
                      - isSandbox
                      - createdAt
                      - updatedAt
                    additionalProperties: false
                  apiKey:
                    type: string
                required:
                  - merchant
                  - apiKey
                additionalProperties: false
        "409":
          description: Merchant already exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  requestId:
                    type: string
                required:
                  - error
                  - requestId
                additionalProperties: false
  /api/v1/merchants/me:
    get:
      tags:
        - Merchants
      summary: Get merchant profile
      security:
        - apiKeyAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Merchant profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  name:
                    type: string
                  email:
                    type: string
                    format: email
                    pattern: ^(?!\.)(?!.*\.\.)([A-Za-z0-9_'+\-\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
                  agentOptIn:
                    type: boolean
                  settlementType:
                    type: string
                  settlementTiming:
                    type: string
                  webhookUrl:
                    anyOf:
                      - type: string
                      - type: "null"
                  webhookSecret:
                    anyOf:
                      - type: string
                      - type: "null"
                  isSandbox:
                    type: boolean
                  createdAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                  updatedAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                required:
                  - id
                  - name
                  - email
                  - agentOptIn
                  - settlementType
                  - settlementTiming
                  - webhookUrl
                  - isSandbox
                  - createdAt
                  - updatedAt
                additionalProperties: false
        "401":
          description: Auth error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  requestId:
                    type: string
                required:
                  - error
                  - requestId
                additionalProperties: false
  /api/v1/merchants/me/config:
    put:
      tags:
        - Merchants
      summary: Update merchant config
      security:
        - apiKeyAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                agentOptIn:
                  type: boolean
                settlementType:
                  type: string
                  enum:
                    - discount
                    - cashback
                    - credit
                settlementTiming:
                  type: string
                  enum:
                    - real_time
                    - daily_batch
                webhookUrl:
                  anyOf:
                    - type: string
                      format: uri
                    - type: "null"
      responses:
        "200":
          description: Updated merchant
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  name:
                    type: string
                  email:
                    type: string
                    format: email
                    pattern: ^(?!\.)(?!.*\.\.)([A-Za-z0-9_'+\-\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\-]*\.)+[A-Za-z]{2,}$
                  agentOptIn:
                    type: boolean
                  settlementType:
                    type: string
                  settlementTiming:
                    type: string
                  webhookUrl:
                    anyOf:
                      - type: string
                      - type: "null"
                  webhookSecret:
                    anyOf:
                      - type: string
                      - type: "null"
                  isSandbox:
                    type: boolean
                  createdAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                  updatedAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                required:
                  - id
                  - name
                  - email
                  - agentOptIn
                  - settlementType
                  - settlementTiming
                  - webhookUrl
                  - isSandbox
                  - createdAt
                  - updatedAt
                additionalProperties: false
        "401":
          description: Auth error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  requestId:
                    type: string
                required:
                  - error
                  - requestId
                additionalProperties: false
  /api/v1/merchants/me/earn-rules:
    get:
      tags:
        - Merchants
      summary: List earn rules
      security:
        - apiKeyAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Earn rules
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                      pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                    merchantId:
                      type: string
                      format: uuid
                      pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                    name:
                      type: string
                    spendThreshold:
                      type: number
                    tokenDenomination:
                      type: number
                    expiryDays:
                      type: integer
                      minimum: -9007199254740991
                      maximum: 9007199254740991
                    categoryRestriction:
                      anyOf:
                        - type: string
                        - type: "null"
                    channelRestriction:
                      anyOf:
                        - type: string
                        - type: "null"
                    stackabilityFlag:
                      type: boolean
                    agentPresentableFlag:
                      type: boolean
                    minimumTransactionFloor:
                      type: number
                    isActive:
                      type: boolean
                    createdAt:
                      type: string
                      format: date-time
                      pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                    updatedAt:
                      type: string
                      format: date-time
                      pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                  required:
                    - id
                    - merchantId
                    - name
                    - spendThreshold
                    - tokenDenomination
                    - expiryDays
                    - categoryRestriction
                    - channelRestriction
                    - stackabilityFlag
                    - agentPresentableFlag
                    - minimumTransactionFloor
                    - isActive
                    - createdAt
                    - updatedAt
                  additionalProperties: false
        "401":
          description: Auth error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  requestId:
                    type: string
                required:
                  - error
                  - requestId
                additionalProperties: false
    post:
      tags:
        - Merchants
      summary: Create earn rule
      security:
        - apiKeyAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                spendThreshold:
                  type: number
                  minimum: 0
                tokenDenomination:
                  type: number
                  exclusiveMinimum: 0
                expiryDays:
                  default: 90
                  type: integer
                  exclusiveMinimum: 0
                  maximum: 9007199254740991
                categoryRestriction:
                  anyOf:
                    - type: string
                    - type: "null"
                channelRestriction:
                  anyOf:
                    - type: string
                    - type: "null"
                stackabilityFlag:
                  default: true
                  type: boolean
                agentPresentableFlag:
                  default: true
                  type: boolean
                minimumTransactionFloor:
                  default: 0
                  type: number
                  minimum: 0
              required:
                - name
                - spendThreshold
                - tokenDenomination
      responses:
        "201":
          description: Earn rule created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  merchantId:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  name:
                    type: string
                  spendThreshold:
                    type: number
                  tokenDenomination:
                    type: number
                  expiryDays:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                  categoryRestriction:
                    anyOf:
                      - type: string
                      - type: "null"
                  channelRestriction:
                    anyOf:
                      - type: string
                      - type: "null"
                  stackabilityFlag:
                    type: boolean
                  agentPresentableFlag:
                    type: boolean
                  minimumTransactionFloor:
                    type: number
                  isActive:
                    type: boolean
                  createdAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                  updatedAt:
                    type: string
                    format: date-time
                    pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                required:
                  - id
                  - merchantId
                  - name
                  - spendThreshold
                  - tokenDenomination
                  - expiryDays
                  - categoryRestriction
                  - channelRestriction
                  - stackabilityFlag
                  - agentPresentableFlag
                  - minimumTransactionFloor
                  - isActive
                  - createdAt
                  - updatedAt
                additionalProperties: false
        "401":
          description: Auth error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  requestId:
                    type: string
                required:
                  - error
                  - requestId
                additionalProperties: false
...
  /api/v1/merchants/me/redemptions:
    get:
      tags:
        - Merchants
      summary: List merchant redemptions
      security:
        - apiKeyAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Merchant redemptions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      format: uuid
                      pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                    tokenId:
                      type: string
                      format: uuid
                      pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                    merchantId:
                      type: string
                      format: uuid
                      pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                    customerId:
                      type: string
                      format: uuid
                      pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                    transactionAmount:
                      type: number
                    tokenDenomination:
                      type: number
                    netValue:
                      type: number
                    agentId:
                      anyOf:
                        - type: string
                        - type: "null"
                    settlementRef:
                      type: string
                    redeemedAt:
                      type: string
                      format: date-time
                      pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                  required:
                    - id
                    - tokenId
                    - merchantId
                    - customerId
                    - transactionAmount
                    - tokenDenomination
                    - netValue
                    - agentId
                    - settlementRef
                    - redeemedAt
                  additionalProperties: false
        "401":
          description: Auth error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      details:
                        type: object
                        propertyNames:
                          type: string
                        additionalProperties: {}
                    required:
                      - code
                      - message
                    additionalProperties: false
                  requestId:
                    type: string
                required:
                  - error
                  - requestId
                additionalProperties: false
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT